// apis/trusted-apps-registry-v4/src/modules/apps/apps.interface.ts
export interface AppLink {
  id: string;
  name: string;
  href: string;
}

export interface AppObject {
  applicationId: string;
  name: string;
}

export interface PublicKeyLink {
  id: string;
  href: string;
}

export interface AuthorizationLink {
  authorizationId: string;
  requesterApplicationName: string;
  href: string;
}

export interface AuthorizationItemObject {
  authorizationId: string;
  authorizedAppName: string;
}

export interface AuthorizationObject {
  applicationId: string;
  authorizedAppId: string;
  name: string;
  authorizedAppName: string;
  iss: string;
  status: number;
}

export interface AuthorizationResponseObject {
  authorizationId: string;
  resourceApplicationId: string;
  requesterApplicationId: string;
  resourceApplicationName: string;
  requesterApplicationName: string;
  iss: string;
  status: string;
}

export interface AppRevocationInfo {
  revokedBy: string;
  notBefore: number;
}

export interface AppResponseObject {
  applicationId: string;
  name: string;
  domain: string;
  administrators: string[];
  publicKeys: string[];
  info: {
    [x: string]: unknown;
  };
  authorizations: AuthorizationResponseObject[];
  revocation: AppRevocationInfo;
}

export interface PublicKeyResponseObject {
  applicationId: string;
  publicKey: string;
  status: string;
}